# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/auth/profile/avatar": {
      "post": {
        "description": "Usuario autenticado sube o reemplaza SU imagen de perfil. Formatos: JPG, PNG, WebP. Tamaño máximo: 2MB. Dimensiones: 800x800px.",
        "operationId": "ProfileController_uploadAvatar",
        "parameters": [],
        "requestBody": {
          "required": true,
          "description": "Subir avatar de usuario (JPG, PNG, WebP)",
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo (.jpg, .jpeg, .png, .gif, .webp, .svg)"
                  }
                },
                "required": [
                  "avatar"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Imagen subida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Subir imagen de perfil (self-service)",
        "tags": [
          "auth-profile"
        ]
      },
      "delete": {
        "description": "Usuario autenticado elimina SU imagen de perfil del storage y retorna el usuario actualizado.",
        "operationId": "ProfileController_deleteAvatar",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Imagen eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Eliminar imagen de perfil (self-service)",
        "tags": [
          "auth-profile"
        ]
      }
    },
    "/api/auth/profile/password": {
      "patch": {
        "description": "Usuario autenticado cambia SU contraseña. Requiere contraseña actual para validación. Revoca todas las sesiones activas y dispositivos confiables por seguridad.",
        "operationId": "ProfileController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Contraseña actualizada. Sesiones cerradas. Usuario debe volver a iniciar sesión.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Contraseña actualizada exitosamente. Se han cerrado todas las sesiones por seguridad. Deberás volver a iniciar sesión."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Contraseña actual incorrecta o nueva contraseña inválida"
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Cambiar contraseña (self-service)",
        "tags": [
          "auth-profile"
        ]
      }
    },
    "/api/auth/profile/2fa/activate": {
      "patch": {
        "description": "Usuario autenticado activa la autenticación de dos factores en SU cuenta. A partir de ahora, deberá ingresar un código enviado por email al iniciar sesión.",
        "operationId": "ProfileController_activate2fa",
        "parameters": [],
        "responses": {
          "200": {
            "description": "2FA activado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Activar 2FA (self-service)",
        "tags": [
          "auth-profile"
        ]
      }
    },
    "/api/auth/profile/2fa/deactivate": {
      "patch": {
        "description": "Usuario autenticado desactiva la autenticación de dos factores en SU cuenta. Ya no se requerirá código de verificación al iniciar sesión.",
        "operationId": "ProfileController_deactivate2fa",
        "parameters": [],
        "responses": {
          "200": {
            "description": "2FA desactivado exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/UserResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Solicitud incorrecta",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "404": {
            "description": "Recurso no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "type": "null",
                          "example": null
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "summary": "Desactivar 2FA (self-service)",
        "tags": [
          "auth-profile"
        ]
      }
    },
...
    "/api/users/{id}/deactivate": {
      "patch": {
        "operationId": "UsersController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del usuario a SUSPENDED y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/activate": {
      "patch": {
        "operationId": "UsersController_activate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del usuario a ACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/2fa/activate": {
      "patch": {
        "operationId": "UsersController_activate2fa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del 2fa de un usuario a ACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/2fa/deactivate": {
      "patch": {
        "operationId": "UsersController_deactivate2fa",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID del recurso",
            "schema": {
              "format": "uuid",
              "example": "123e4567-e89b-12d3-a456-426614174000",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Cambia el estado del 2fa de un usuario a INACTIVE y retorna un mensaje de confirmación.",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessResponseDto"
                    },
                    {
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/MessageResponseDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "ATR API",
    "description": "API de auditorías, plantillas, frameworks de madurez y gestión de usuarios",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "SuccessResponseDto": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Indica si la operación fue exitosa",
            "example": true,
            "default": true
          },
          "statusCode": {
            "type": "number",
            "description": "Código de estado HTTP",
            "example": 200,
            "enum": [
              200,
              201,
              204
            ]
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación. Se genera automáticamente según el método HTTP o se personaliza con @ResponseMessage",
            "example": "Operación exitosa"
          },
          "data": {
            "type": "object",
            "description": "Datos de la operación. Será null si el endpoint no retorna datos.",
            "nullable": true
          },
          "timestamp": {
            "type": "string",
            "description": "Timestamp ISO 8601 de cuando se generó la respuesta",
            "example": "2024-01-30T10:30:00.000Z"
          }
        },
        "required": [
          "success",
          "statusCode",
          "message",
          "timestamp"
        ]
      },
...
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "ID único del usuario",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "names": {
            "type": "string",
            "description": "Nombres del usuario",
            "example": "Juan Carlos"
          },
          "lastNames": {
            "type": "string",
            "description": "Apellidos del usuario",
            "example": "Pérez García"
          },
          "email": {
            "type": "string",
            "description": "Email del usuario",
            "example": "juan.perez@example.com"
          },
          "username": {
            "type": "string",
            "description": "Nombre de usuario único",
            "example": "juan_perez"
          },
          "ci": {
            "type": "string",
            "description": "Cédula de identidad",
            "example": "1234567-8"
          },
          "phone": {
            "type": "object",
            "description": "Teléfono de contacto",
            "example": "+591 2 234 5678",
            "nullable": true
          },
          "address": {
            "type": "object",
            "description": "Dirección del usuario",
            "example": "Av. 6 de Agosto #123",
            "nullable": true
          },
          "image": {
            "type": "object",
            "description": "URL de la imagen de perfil",
            "example": "https://example.com/images/profile.jpg",
            "nullable": true
          },
          "isActive": {
            "type": "boolean",
            "description": "Estado del usuario",
            "example": true
          },
          "isTemporaryPassword": {
            "type": "boolean",
            "description": "Indica si el usuario debe cambiar su contraseña temporal",
            "example": true
          },
          "firstLoginAt": {
            "type": "object",
            "description": "Fecha y hora del primer login del usuario",
            "example": "2024-01-15T10:30:00.000Z",
            "nullable": true
          },
          "isTwoFactorEnabled": {
            "type": "boolean",
            "description": "Indica si el usuario tiene 2FA habilitado",
            "example": false
          },
          "roles": {
            "type": "array",
            "description": "Roles asignados al usuario",
            "example": [
              "auditor"
            ],
            "items": {
              "type": "string",
              "enum": [
                "admin",
                "gerente",
                "auditor",
                "cliente"
              ]
            }
          },
          "organizationId": {
            "type": "string",
            "description": "ID de la organización a la que pertenece",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "organizationName": {
            "type": "string",
            "description": "Nombre de la organización",
            "example": "Empresa S.A."
          },
          "organizationImage": {
            "type": "object",
            "description": "URL de la imagen de la organización",
            "example": "https://example.com/images/org-logo.jpg",
            "nullable": true
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de creación del usuario",
            "example": "2024-01-15T10:30:00.000Z"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Fecha de última actualización",
            "example": "2024-01-15T10:30:00.000Z"
          }
        },
        "required": [
          "id",
          "names",
          "lastNames",
          "email",
          "username",
          "ci",
          "isActive",
          "isTemporaryPassword",
          "isTwoFactorEnabled",
          "roles",
          "organizationId",
          "organizationName",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "MessageResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo de la operación",
            "example": "Operación completada exitosamente"
          }
        },
        "required": [
          "message"
        ]
      },
...
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "description": "Contraseña actual del usuario",
            "example": "OldPassword123!",
            "minLength": 8,
            "maxLength": 100
          },
          "newPassword": {
            "type": "string",
            "description": "Nueva contraseña",
            "example": "NewPassword123!",
            "minLength": 8,
            "maxLength": 100
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  }
}